+++ ui/ui-components/components/badges/appBadge.tsx
@@ -1,5 +1,5 @@
 import { Badge } from "react-bootstrap";
-import EnterpriseLink from "../enterpriseLink";
+import EnterpriseLink from "../grouparooLink";
 
 export default function AppBadge({
   appName,
